test: OverflowMenu integration
